separate components and actions hook
