+++ extensions/reclaim-ai/src/scheduling-links.tsx
@@ -1,68 +1,68 @@
-import { Action, ActionPanel, Icon, List, Toast, open, showToast } from "@raycast/api";
+import { Icon, List } from "@raycast/api";
 import { useMemo, useState } from "react";
 import { useSchedulingLinks } from "./hooks/useSchedulingLinks";
 import { useUser } from "./hooks/useUser";
-import { SchedulingLink } from "./types/scheduling-link";
-import { axiosPromiseData, fetcher } from "./utils/axiosPromise";
+import { ListItemDetailMetadataField } from "./components/ListItemDetailMetadataField";
+import { SchedulingLinkActionPanel } from "./components/SchedulingLinkActionPanel";
 import { resolveTimePolicy } from "./utils/time-policy";
 
-const SLActions = ({ link }: { link: SchedulingLink }) => {
-  const url = `https://app.reclaim.ai/m/${link.pageSlug}/${link.slug}`;
-  const { currentUser } = useUser();
+// const SLActions = ({ link }: { link: SchedulingLink }) => {
+//   const url = `https://app.reclaim.ai/m/${link.pageSlug}/${link.slug}`;
+//   const { currentUser } = useUser();
 
-  const shareTimesEnabled = useMemo(() => {
-    return !!currentUser?.features.schedulingLinks.shareTimesEnabled;
-  }, [currentUser]);
+//   const shareTimesEnabled = useMemo(() => {
+//     return !!currentUser?.features.schedulingLinks.shareTimesEnabled;
+//   }, [currentUser]);
 
-  const createOneOffLink = async () => {
-    const [oneOff, error] = await axiosPromiseData<SchedulingLink>(
-      fetcher("/scheduling-link/derivative", {
-        method: "POST",
-        data: {
-          parentId: link.id,
-        },
-      })
-    );
+//   const createOneOffLink = async () => {
+//     const [oneOff, error] = await axiosPromiseData<SchedulingLink>(
+//       fetcher("/scheduling-link/derivative", {
+//         method: "POST",
+//         data: {
+//           parentId: link.id,
+//         },
+//       })
+//     );
 
-    if (!error && oneOff) {
-      open(`https://app.reclaim.ai/scheduling-links?personalize=${oneOff.id}`);
-    } else {
-      await showToast({
-        style: Toast.Style.Failure,
-        title: "Error creating one-off link. Try using the reclaim.ai web app.",
-      });
-    }
-  };
+//     if (!error && oneOff) {
+//       open(`https://app.reclaim.ai/scheduling-links?personalize=${oneOff.id}`);
+//     } else {
+//       await showToast({
+//         style: Toast.Style.Failure,
+//         title: "Error creating one-off link. Try using the reclaim.ai web app.",
+//       });
+//     }
+//   };
 
-  const createShareLink = async () => {
-    open(`https://app.reclaim.ai/quick-forms/scheduling-links/${link.id}/available-times`);
-  };
+//   const createShareLink = async () => {
+//     open(`https://app.reclaim.ai/quick-forms/scheduling-links/${link.id}/available-times`);
+//   };
 
-  return (
-    <ActionPanel>
-      <Action.CopyToClipboard title="Copy Link to Clipboard" content={url} />
-      {shareTimesEnabled && <Action icon={Icon.AddPerson} title="Personalize and Share" onAction={createShareLink} />}
-      {!shareTimesEnabled && <Action icon={Icon.AddPerson} title="Create One Off Link" onAction={createOneOffLink} />}
-      <Action.Open title="Open in Browser" target={url} />
-    </ActionPanel>
-  );
-};
+//   return (
+//     <ActionPanel>
+//       <Action.CopyToClipboard title="Copy Link to Clipboard" content={url} />
+//       {shareTimesEnabled && <Action icon={Icon.AddPerson} title="Personalize and Share" onAction={createShareLink} />}
+//       {!shareTimesEnabled && <Action icon={Icon.AddPerson} title="Create One Off Link" onAction={createOneOffLink} />}
+//       <Action.Open title="Open in Browser" target={url} />
+//     </ActionPanel>
+//   );
+// };
 
-const ListDetailMetadataField = ({
-  title,
-  value,
-  description,
-}: {
-  title: string;
-  value: string;
-  description?: string;
-}) => (
-  <>
-    <List.Item.Detail.Metadata.Label title={title} />
-    <List.Item.Detail.Metadata.Label title={value} text={description} />
-    <List.Item.Detail.Metadata.Separator />
-  </>
-);
+// const ListDetailMetadataField = ({
+//   title,
+//   value,
+//   description,
+// }: {
+//   title: string;
+//   value: string;
+//   description?: string;
+// }) => (
+//   <>
+//     <List.Item.Detail.Metadata.Label title={title} />
+//     <List.Item.Detail.Metadata.Label title={value} text={description} />
+//     <List.Item.Detail.Metadata.Separator />
+//   </>
+// );
 
 export default function Command() {
   const [searchText, setSearchText] = useState("");
@@ -122,18 +122,18 @@ export default function Command() {
 
                           return (
                             <List.Item.Detail.Metadata>
-                              <ListDetailMetadataField title="Hours:" value={hours} />
-                              <ListDetailMetadataField title="Organizers:" value={organizers} />
-                              <ListDetailMetadataField title="Duration:" value={duration} />
+                              <ListItemDetailMetadataField title="Hours:" value={hours} />
+                              <ListItemDetailMetadataField title="Organizers:" value={organizers} />
+                              <ListItemDetailMetadataField title="Duration:" value={duration} />
                               {sl.priority === "HIGH" && (
-                                <ListDetailMetadataField title="Priority:" value="High Priority" />
+                                <ListItemDetailMetadataField title="Priority:" value="High Priority" />
                               )}
                             </List.Item.Detail.Metadata>
                           );
                         })()}
                       />
                     }
-                    actions={<SLActions link={sl} />}
+                    actions={<SchedulingLinkActionPanel link={sl} />}
                   />
                 ))}
             </List.Section>
